requirements.txt
langchain==0.3.25
langchain-core==0.3.63
langchain-community==0.3.23
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
openai==1.75.0
chromadb==0.6.3
sentence-transformers==3.4.1
python-dotenv==1.1.0
PyGithub==2.6.1
aiohttp
